Added first working version
At this time, most of the first project has been transferred from old
machine to new machine, and reconfigured in MyInventoryApp instead of
LoginApp.  Add product function is working, but update and delete only
work on the webpage itself, but do not make any changes inside the
actual PhpMyAdmin database.  Confused about this.  Needs to be fixed.

Next Steps: Switch over to MVC architecture in order to implement Spring
Security.