Enables Controller Functionality, Adds "Security"
The Controller is now functional with a GET and POST method.
Successfully instantiates Users to compare with Hibernate database and
to pass fields to other web pages.  A dummy version of security is
implemented.  There is no more hyperlink straight to the inventory page.
Now you must login before being redirected to inventory.